added in road generation, scale is controlled by the highway config. Its only generated at startup so I kept it simple.

+++ Ported/HighwayRacers-gr3/Assets/Scripts/Authoring/HighwayAuthoring.cs
@@ -4,7 +4,7 @@ class HighwayConfigAuthoring : UnityEngine.MonoBehaviour
 {
     public UnityEngine.GameObject StraightRoadPrefab;
     public UnityEngine.GameObject CurvedRoadPrefab;
-    public float InsideLaneLength;
+    public int InsideLaneLength;
 }
 
 class HighwayConfigBaker : Baker<HighwayConfigAuthoring>
